init `contrib.models`
