+++ kornia/contrib/models/__init__.py
@@ -1,0 +1,2 @@
+from kornia.contrib.models.prompters import Prompts
+from kornia.contrib.models.structures import SegmentationResults
