+++ kornia/contrib/models/prompters/__init__.py
@@ -1,0 +1,2 @@
+from kornia.contrib.models.prompters.base import PrompterModelBase
+from kornia.contrib.models.prompters.structures import Prompts
